Added FixeModelEntityItem and FormFactory
This allows the creation of a EntityEditor window which doesn't
have the ability of switching between existing models of the same
Entity. Such a case are the models of the Material. We don't
want to switch between a DisneyMaterial and any other type of material.

Most of the code is written by Esteban Tovagliari.

Signed-off-by: Marius Avram <marius.avram1309@gmail.com>