DisneyMaterialCustomUI: use QFormLayout instead of QVBoxLayout
This is the same kind of layout used in EntityEditor.
QVBoxLayout was used in case the Expression Editor was to be
drawn in the same window with the DisneyMaterial editor.

But because it opens in a new dialog there is no need for it.
Also this eliminates the need for a fixed sized QLabel, which
is not according to the variable sized text size from the interface.

Signed-off-by: Marius Avram <marius.avram1309@gmail.com>