Preserve newlines in parameter value
Newlines were not loaded properly for a value containing them. The cause
is the transcode() function from xerces which ignores them.

To solve the issue, newline characters are replaced with explicit \n when
saving and converted back again to newlines when loading a scene.

This is required for disney expressions which is a script found on multiple
lines.

Signed-off-by: Marius Avram <marius.avram1309@gmail.com>