Save folded disney material layer state between sessions
When a project will be saved the folding state will also be saved
in the project file as a simple parameter with the true or false
value.

The next time the project will be opened the parameter will be used
to fold the layers the same way as they were before appleseed.studio
was closed.

Signed-off-by: Marius Avram <marius.avram1309@gmail.com>